+++ scripts/devops_tasks/test_run_samples.py
@@ -138,7 +138,11 @@ IGNORED_SAMPLES = {
         "sample_analyze_healthcare_entities_with_cancellation_async.py",
         "sample_abstract_summary.py",
         "sample_abstract_summary_async.py",
-    ]
+    ],
+    "azure-ai-language-conversations": [
+        "sample_assign_deployment_resources.py",
+        "sample_assign_deployment_resources_async.py",
+    ],
 }
 
 def run_check_call_with_timeout(