+++ sdk/ml/azure-ai-ml/azure/ai/ml/_schema/compute/compute_instance.py
@@ -66,7 +66,7 @@ class ComputeInstanceSchema(ComputeSchema):
     )
     schedules = NestedField(ComputeSchedulesSchema)
     identity = ExperimentalField(NestedField(IdentitySchema))
-    idle_time_before_shutdown = ExperimentalField(fields.Str())
+    idle_time_before_shutdown_minutes = ExperimentalField(fields.Int())
     setup_scripts = ExperimentalField(NestedField(SetupScriptsSchema))
     os_image_metadata = ExperimentalField(
         NestedField(OsImageMetadataSchema, dump_only=True)