+++ sdk/ml/azure-ai-ml/tests/compute/unittests/test_compute_entity.py
@@ -191,12 +191,12 @@ class TestComputeEntity:
             "tests/test_configs/compute/compute-ci-schedules.yaml"
         )
         assert len(compute_instance.schedules.compute_start_stop) == 2
-        assert compute_instance.idle_time_before_shutdown == "PT15M"
+        assert compute_instance.idle_time_before_shutdown_minutes == 15
 
         compute_resource = compute_instance._to_rest_object()
-        compute_instance2: ComputeInstance = ComputeInstance._load_from_rest(
-            compute_resource
-        )
+        assert compute_resource.properties.properties.idle_time_before_shutdown == f"PT{compute_instance.idle_time_before_shutdown_minutes}M"
+        
+        compute_instance2: ComputeInstance = ComputeInstance._load_from_rest(compute_resource)
         assert len(compute_instance2.schedules.compute_start_stop) == 2
         assert compute_instance2.schedules.compute_start_stop[0].action == "stop"
         assert compute_instance2.schedules.compute_start_stop[0].trigger.type == "Cron"
@@ -230,10 +230,8 @@ class TestComputeEntity:
             == "week"
         )
         assert compute_instance2.schedules.compute_start_stop[1].trigger.interval == 1
-        assert (
-            compute_instance2.schedules.compute_start_stop[1].trigger.schedule
-            is not None
-        )
+        assert compute_instance2.schedules.compute_start_stop[1].trigger.schedule is not None
+        assert compute_instance2.idle_time_before_shutdown_minutes == compute_instance.idle_time_before_shutdown_minutes
 
     def test_compute_instance_setup_scripts_from_yaml(self):
         loaded_instance: ComputeInstance = load_compute(